feat: allow uploading of files

+++ library/boa-core/src/packets/server/process/event.rs
@@ -3,6 +3,6 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum ProcessEventPacket {
     Started,
-    Finished { exit_code: i32 },
+    Finished { exit_code: i64 },
     TimedOut,
 }
